+++ memote/__init__.py
@@ -25,6 +25,7 @@ genome-scale metabolic models.
 from __future__ import absolute_import
 
 from memote.version_info import show_versions
+from memote.suite.api import *
 
 __author__ = "Moritz E. Beber"
 __email__ = "morbeb@biosustain.dtu.dk"